gh: remove CODEOWNERS
Code owners is a concept in GH that allows specifying maintainers of a package/directory or otherwise a set of files.
This is quite useful in big repositories where many ppl contribute to formalize responsibility.
In this repository it serves no role, more info can be found in [gh code owners documentation](https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners#about-code-owners).