makefile: dev, replace air with properly configured make target
- Remove air from install-dependencies
- Remove air config file
- Change phony dev-build to (real) forwarder.race
  - it depends on relevant go files
  - linking flags are removed:
    - static
      - statically linking binary takes more time - not needed for devel
      - -race makes -static irrelevant because the race detector needs to be linked dynamically
  - version.* - this is set by goreleaser when releasing, not needed for development, removes ambiguity
- Change dev target, it depends on and runs forwarder.race

Case against air, and hot reload in general.

Hot reload is common practice in web development.
The idea is that one is working on a web page and tests the whole stack all the time without a need to manually update backend and redeploy js/html.

This is not something that you would want when working on a proxy.
When running a binary and proxying some traffic for a real life test the last thing you want is a casual restart.
Also, time spent on manually tweaking code while proxying would be better spent writing a test case that implements the scenario you want to reproduce.