credentials: refactor userInfoMatcher into CredentialsMatcher
It's exported, the intention is to make it a HTTTProxy dependency not something it creates internally.
New type HostPortUser is added, and parsing is moved to separate function.
Constructor accepts a list of HostPortUser.
normalizeURLScheme is removed in favour of adding "http://" prefix to value so that we can use the url Parser.