cmd/forwarder: add httpbin command implementing elements of httpbin.org API
The implemented endpoints are:
- `/basic-auth/{user}/{passwd}`
- `/delay/{seconds}`
- `/status/{code}`
- `/stream-bytes/{bytes}`

Using this approach we reuse the server code for HTTPS.
We do not need third party containers and have full control over performance of the site in test in case we want to benchmark the proxy.
Also, we have flexibility to extend it with web sockets etc.