martian: remove redundant buffers in copySync
As demonstrated in previous commit the buffers get in a way of TLS handshake in HTTP CONNECT (without MITM) workflow.
This patch removes the double buffering in this case.
The io.Copy() internally also uses a buffer, the default buffer size is `32 * 1024`.
In case of plain TCP connections (no TLS) on both ends this patch enables SPLICE optimization in io.Copy() [1].

[1] https://github.com/golang/go/issues/10948