martian: perf: Do not spawn a go routine to read request
Without this patch http.ReadRequest() is called in an anonymous function in a go routine.
This is to check if closing channel was not closed while reading.
This requires allocating memory for the anonymous function, channels and waiting on a select (epoll?) all in a hot path.

This patch brings http.ReadRequest() call to the connection handling go routine and uses select/default technique to check if closing channel was closed.