martian(fifo): add ImmutableGroup
This patch introduces group struct which will be the core of all fifo.Groups

ImmutableGroup is a martian.RequestResponseModifier that maintains lists of request and response modifiers executed on a first-in, first-out basis.
ImmutableGroup can be constructed only from a fifo.Group, since it cannot be modified, it has no locks.