http_proxy: don't use Host header in FindProxyForURL() host parameter
The Mozilla docs state that the host parameter is extracted from URL [1].

host
The hostname extracted from the URL. This is only for convenience; it is the same string as between :// and the first : or / after that. The port number is not included in this parameter. It can be extracted from the URL when necessary.

The Host header may contain a port, this is unexpected and may break PAC functions.

[1] https://developer.mozilla.org/en-US/docs/Web/HTTP/Proxy_servers_and_tunneling/Proxy_Auto-Configuration_PAC_file#parameters

Fixes #184