martian: handle CONNECT before modifying URL scheme
Connect request should not have a scheme.

CONNECT uses a special form of request target, unique to this method, consisting of only the host and port number of the tunnel destination, separated by a colon. There is no default port; a client MUST send the port number even if the CONNECT request is based on a URI reference that contains an authority component with an elided port (Section 4.1). For example,

CONNECT server.example.com:80 HTTP/1.1
Host: server.example.com

https://httpwg.org/specs/rfc9110.html#CONNECT