martian: an http.Handler implementation
This patch provides a proxyHandler type that wraps a Proxy and implements http.Handler.
It is not intended to replace the existing Proxy, but to provide a more convenient way to use it where a http.Handler is required.
One such example is HTTP/2 support, where a proxyHandler can be used as a handler for http2.Server.
The handler works with both HTTP/1 and HTTP/2 requests.

The implementation is based on httputil.ReverseProxy [1],
but it keeps the Martian Proxy structure and uses its unexported methods.
Unlike httputil.ReverseProxy, proxyHandler does not expose any additional configuration, and some assumptions are made.

- We do not use the httputil.ReverseProxy's Director, but we use Martian's Request modifiers.
- We use a copy buffer pool by default.
- We either manually flush after every write or not at all, this is the same as setting FlushInterval to 0.
- We always flush after the response headers are written, this is for compatibility with the existing Proxy.
- The above has a side effect of preventing net/http from calculating the length for short bodies and adding a Content-Length.

The MITM functionality is not supported, but it could be added in the future.

Prior art:
- httputil.ReverseProxy [1]
- Caddy's forwardproxy [2]
- elazarl/goproxy [3]

[1] https://pkg.go.dev/net/http/httputil#ReverseProxy
[2] https://github.com/caddyserver/forwardproxy
[3] https://github.com/elazarl/goproxy