config,dns: harden DNS server address parsing
This changes the type used for DNS servers to netip.AddrPort.
It saves as from many unnecessary validations and improves the error reporting.

Sample errors:
- invalid argument "1.1.1.1.1" for "-n, --dns-server" flag: IP: ParseAddr("1.1.1.1.1"): IPv4 address too long
- invalid argument "saucelabs.com" for "-n, --dns-server" flag: IP: ParseAddr("saucelabs.com"): unexpected character (at "saucelabs.com")
- invalid argument "1.1.1.1:444444" for "-n, --dns-server" flag: port: strconv.ParseUint: parsing "444444": value out of range

This patch also improves port validation by using proper types (and strconv.ParseUint(port, 10, 16)).

Last but not least the help flag descriptions are extended.
This is a new format that would be applied to all flags.