utils/cobrautil/templates: support custom value information
Enable displaying things like '-x, --proxy [protocol://]host[:port]' without littering the usage.

Normally you have to quote (``) a string in usage to print it next to name.
This adds possibility to add type value information as usage prefix that is consumed and removed from the output.
In addition, the cobra default type name is wrapped into <X> for coherence.
This system is largely inspired by curl but can be found in man pages for other UNIX tools too.