header: a package to add and remove headers
The following syntax (and actions) is supported:
- "<name>: <value>" to add a header,
- "<name>;" to set a header to empty,
- "-<name>" to remove a header,
- "-<name>*" to remove a header by prefix.

This syntax is an extension of the curl -H, --header flag syntax,
extended with the ability to remove the headers.

This approach has advantages over a flag per action approach, like one found in ngrok.

    --request-header-add strings       header key:value to add to request
    --request-header-remove strings    header field to remove from request if present
    --response-header-add strings      header key:value to add to response
    --response-header-remove strings   header field to remove from response if present

With this approach the command help can be simpler and compatible with curl.
Also, we control the order of execution.
We can mix and match additions and deletions from different sources.