e2e: add DNS setup and test
The setup use custom docker network called "forwarder-e2e_default", which might be misleading, as this is not the default network.
This in on purpose to use Makefile target "make test" without excessive changes to the Makefile itself.

The setups consist of bind9 custom dns server that resolves "httpbin.local" hostname to ipv4 address.