gh: add action to assert linked modules changes
Forwarder uses go:linkname directive. It may lead to unsafe behavior if the linked code changes.
This workflow performs a primitive check by calculating the checksums of linked modules.
If the checksums do not match, a user is forced to manually check the changes and update the checksums.

The action runs on '.version' file changes.