e2e: run the e2e test on gcr.io/distroless
The ubuntu:latest image does not have the updated root certificates and required apt-get update to fix it.
This prevents removal of insecure TLS mode for the test binary.

The gcr.io/distroless is more practical and lighter.
It is the same image we base Forwarder image on.
In CI it does not require downloading any additional image.