http_proxy: deny domains
Deny domains specify proxying to which domains is disabled.
It is configured via two lists of regular expressions - Include and Exclude lists.
When a hostname matches at least one expression in the Include list and does not match
any expression in the Exclude list the request is denied.

When only the Exclude rules are specified, all the requests that does not match any Exclude rules are denied.

Fixes #305