dialvia(http): read CONNECT response by byte
Add byteReader, a reader that reads only one byte at a time.
Wrap connection in byteReader when passing to bufio.NewReaderSize() - bufio is required due to http.ReadResponse() api.

Read buffer size is set to 1. It's actually reset to 16 in bufio Reader constructor but since we use byteReader the buffer is never filled with more that a single byte.

Fixes #616