martian: add initial version of ProxyTrace type
ProxyTrace is modeled after net/http/httptrace package, but it's intended to be attached to Proxy not individual requests - hence the API is slightly different.

It's possible to replicate httptrace.ClientTrace API in ProxyTrace but it's not a goal at the moment.

It's added as means of notifying users that request is fully processed and response is written but can be easily extended to other usecases.