utils/httpbin: set identity transfer encoding for sse
From http.Server:

HTTP/1.1 or greater: Transfer-Encoding has been set to identity, and no
content-length has been provided. The connection must be closed after the
reply is written, and no chunking is to be done. This is the setup
recommended in the Server-Sent Events candidate recommendation 11,
section 8.
if hasTE && te == "identity" {
	cw.chunking = false
	w.closeAfterReply = true
	delHeader("Transfer-Encoding")
} else {
	// HTTP/1.1 or greater: use chunked transfer encoding
	// to avoid closing the connection at EOF.
	cw.chunking = true
	setHeader.transferEncoding = "chunked"
	if hasTE && te == "chunked" {
		// We will send the chunked Transfer-Encoding header later.
		delHeader("Transfer-Encoding")
	}
}