utils/compose/network: remove ip range from IPAM config
In Docker Compose network settings, the ip_range field within the ipam.config section specifies a range of IP addresses
from the defined subnet that can be allocated to containers.
If it is left empty it means no specific range is set, and Docker can allocate any address from the entire subnet.

We don't need ip range for the sake of e2e tests. That would keep the compose as simple as possible.