utils/connfix: utility to simplify wrapping net.Conn without butchering its full interface
When embedding net.Conn (or any interface) the Go compiler looses track of methods provided by implementations other than what the interface specifies.

Sadly we cannot use generics to solve that. It would be best to have

type FooConn[T net.Conn] struct {
    T
}

BUT the compiler claims "Embedded type cannot be a type parameter".

This package fixes that for a very narrow, but important, usecase of overriding any of the following functions:
 - ReadFrom(r io.Reader) (int64, error)
 - WriteTo(w io.Writer) (int64, error) {
 - CloseWrite()

Thanks to that the outer net.Conn can provide a single implementation and persist the original interface.